organize earn components into subdirectory
