+++ packages/demo/frontend/src/components/earn/EarnWithDynamicWallet.tsx
