+++ packages/demo/frontend/src/components/earn/EarnWithFrontendWallet.tsx
