+++ packages/demo/frontend/src/components/earn/EarnWithPrivyServerWallet.tsx
