+++ packages/demo/frontend/src/components/earn/Info.tsx
