+++ packages/demo/frontend/src/components/earn/LoginWithDynamic.tsx
