+++ packages/demo/frontend/src/components/earn/LoginWithPrivy.tsx
