+++ packages/demo/frontend/src/components/earn/Shimmer.tsx
