+++ packages/demo/frontend/src/components/earn/WalletAddressDisplay.tsx
