+++ packages/demo/frontend/src/pages/EarnPage.tsx
@@ -2,12 +2,12 @@ import { useSearchParams } from 'react-router-dom'
 import { PrivyProvider } from '../providers/PrivyProvider'
 import { ActivityLogProvider } from '../providers/ActivityLogProvider'
 import { DynamicProvider } from '@/providers/DynamicProvider'
-import { EarnWithDynamicWallet } from '@/components/EarnWithDynamicWallet'
-import { EarnWithTurnkeyWallet } from '@/components/EarnWithTurnkeyWallet'
-import { EarnWithPrivyServerWallet } from '@/components/EarnWithPrivyServerWallet'
+import { EarnWithDynamicWallet } from '@/components/earn/EarnWithDynamicWallet'
+import { EarnWithTurnkeyWallet } from '@/components/earn/EarnWithTurnkeyWallet'
+import { EarnWithPrivyServerWallet } from '@/components/earn/EarnWithPrivyServerWallet'
 import { WALLET_PROVIDERS } from '@/constants/walletProviders'
 import { TurnkeyProvider } from '../providers/TurnkeyProvider'
-import { WelcomeWalletPicker } from '@/components/WelcomeWalletPicker'
+import { WelcomeWalletPicker } from '@/components/earn/WelcomeWalletPicker'
 
 /**
  * Earn page that renders different wallet provider implementations
